Moved intro/error style in component
